refactor: move partial lock helpers out of cli
